camera movement improved, movement avoided out of world limits
the camera advance smoothly forward the level. Ship follows the movement in case of not moving forward when it tops the right side of the window. The ship can not move outside the world´s limits.
The code that should codify the map collisions was implemented, but it still not working.